refactor: extract cell styles
